Split MCP stdio public contracts
