+++ Sources/QuillCodeTools/MCPStdioProber.swift
@@ -6,81 +6,6 @@ import Darwin
 import Glibc
 #endif
 
-public struct MCPServerProbeResult: Sendable, Hashable {
-    public var protocolVersion: String?
-    public var serverName: String?
-    public var serverVersion: String?
-    public var toolDescriptors: [MCPToolDescriptor]
-    public var toolNames: [String]
-    public var resourceNames: [String]
-    public var resourceURIs: [String]
-    public var promptNames: [String]
-
-    public init(
-        protocolVersion: String? = nil,
-        serverName: String? = nil,
-        serverVersion: String? = nil,
-        toolDescriptors: [MCPToolDescriptor] = [],
-        toolNames: [String] = [],
-        resourceNames: [String] = [],
-        resourceURIs: [String] = [],
-        promptNames: [String] = []
-    ) {
-        self.protocolVersion = protocolVersion
-        self.serverName = serverName
-        self.serverVersion = serverVersion
-        self.toolDescriptors = toolDescriptors.isEmpty
-            ? toolNames.map { MCPToolDescriptor(name: $0) }
-            : toolDescriptors
-        self.toolNames = toolNames.isEmpty
-            ? self.toolDescriptors.map(\.name)
-            : toolNames
-        self.resourceNames = resourceNames
-        self.resourceURIs = resourceURIs
-        self.promptNames = promptNames
-    }
-}
-
-public struct MCPToolDescriptor: Codable, Sendable, Hashable, Identifiable {
-    public var id: String { name }
-    public var name: String
-    public var description: String
-    public var requiredArguments: [String]
-    public var optionalArguments: [String]
-    public var schemaSummary: String
-
-    public init(
-        name: String,
-        description: String = "",
-        requiredArguments: [String] = [],
-        optionalArguments: [String] = [],
-        schemaSummary: String = ""
-    ) {
-        self.name = name
-        self.description = description
-        self.requiredArguments = requiredArguments
-        self.optionalArguments = optionalArguments
-        self.schemaSummary = schemaSummary
-    }
-}
-
-public enum MCPProbeError: LocalizedError, Equatable {
-    case invalidMessage(String)
-    case responseError(String)
-    case timeout(String)
-
-    public var errorDescription: String? {
-        switch self {
-        case .invalidMessage(let message):
-            return message
-        case .responseError(let message):
-            return message
-        case .timeout(let message):
-            return message
-        }
-    }
-}
-
 public final class MCPStdioProber: @unchecked Sendable {
     private let standardInput: FileHandle
     private let standardOutput: FileHandle
@@ -333,30 +258,4 @@ public final class MCPStdioProber: @unchecked Sendable {
         return Data(bytes.prefix(byteCount))
     }
 
-}
-
-public extension ToolDefinition {
-    static let mcpCall = ToolDefinition(
-        name: "host.mcp.call",
-        description: "Call a tool on a verified project-local MCP stdio server. Use only server IDs and tool names listed in the description supplied by QuillCode.",
-        parametersJSON: #"{"type":"object","required":["serverID","toolName"],"properties":{"serverID":{"type":"string"},"toolName":{"type":"string"},"arguments":{"type":"object"},"argumentsJSON":{"type":"string","description":"JSON object string for tool arguments when object arguments are not convenient."}}}"#,
-        host: .mcp,
-        risk: .append
-    )
-
-    static let mcpReadResource = ToolDefinition(
-        name: "host.mcp.resource.read",
-        description: "Read an advertised resource from a verified project-local MCP stdio server. Use only server IDs and resource names or URIs listed in the description supplied by QuillCode.",
-        parametersJSON: #"{"type":"object","required":["serverID"],"properties":{"serverID":{"type":"string"},"resourceURI":{"type":"string","description":"Advertised MCP resource URI."},"uri":{"type":"string","description":"Alias for resourceURI."},"resourceName":{"type":"string","description":"Advertised resource display name when the URI is not convenient."},"name":{"type":"string","description":"Alias for resourceName."}}}"#,
-        host: .mcp,
-        risk: .read
-    )
-
-    static let mcpGetPrompt = ToolDefinition(
-        name: "host.mcp.prompt.get",
-        description: "Get an advertised prompt from a verified project-local MCP stdio server. Use only server IDs and prompt names listed in the description supplied by QuillCode.",
-        parametersJSON: #"{"type":"object","required":["serverID","promptName"],"properties":{"serverID":{"type":"string"},"promptName":{"type":"string"},"name":{"type":"string","description":"Alias for promptName."},"arguments":{"type":"object"},"argumentsJSON":{"type":"string","description":"JSON object string for prompt arguments when object arguments are not convenient."}}}"#,
-        host: .mcp,
-        risk: .read
-    )
 }
