+++ Tests/QuillCodeParityTests/ParityGateTests.swift
@@ -1537,14 +1537,21 @@ final class ParityGateTests: XCTestCase {
         let proberText = try Self.toolsSourceText(named: "MCPStdioProber.swift")
         let codecText = try Self.toolsSourceText(named: "MCPStdioMessageCodec.swift")
         let mapperText = try Self.toolsSourceText(named: "MCPStdioResultMapper.swift")
+        let modelsText = try Self.toolsSourceText(named: "MCPStdioModels.swift")
+        let definitionsText = try Self.toolsSourceText(named: "MCPToolDefinitions.swift")
 
         XCTAssertTrue(proberText.contains("public final class MCPStdioProber"), "MCP stdio session orchestration should remain in the prober.")
         XCTAssertTrue(codecText.contains("public enum MCPStdioMessageCodec"), "MCP Content-Length framing should live in a focused codec.")
         XCTAssertTrue(mapperText.contains("enum MCPStdioResultMapper"), "MCP result mapping should live in a focused mapper.")
+        XCTAssertTrue(modelsText.contains("public struct MCPServerProbeResult"), "MCP probe result models should live outside the stdio prober.")
+        XCTAssertTrue(modelsText.contains("public enum MCPProbeError"), "MCP probe errors should live with the public stdio models.")
+        XCTAssertTrue(definitionsText.contains("static let mcpCall"), "MCP tool definitions should live outside the stdio prober.")
         XCTAssertTrue(proberText.contains("MCPStdioMessageCodec.encodeJSONObject"), "MCP prober should delegate outbound framing to the codec.")
         XCTAssertTrue(proberText.contains("MCPStdioResultMapper.toolDescriptors"), "MCP prober should delegate tool schema summaries to the mapper.")
         XCTAssertTrue(proberText.contains("MCPStdioResultMapper.toolResult"), "MCP prober should delegate tool result formatting to the mapper.")
         XCTAssertFalse(proberText.contains("public enum MCPStdioMessageCodec"), "MCP prober should not own stdio frame parsing.")
+        XCTAssertFalse(proberText.contains("public struct MCPServerProbeResult"), "MCP prober should not own public probe models.")
+        XCTAssertFalse(proberText.contains("public extension ToolDefinition"), "MCP prober should not own static tool definitions.")
         XCTAssertFalse(proberText.contains("private static func schemaArguments"), "MCP prober should not own JSON schema summary formatting.")
         XCTAssertFalse(proberText.contains("private static func toolResult"), "MCP prober should not own ToolResult conversion.")
         XCTAssertFalse(proberText.contains("private static func promptMessageContent"), "MCP prober should not own prompt content flattening.")
